third_party/blink/web_tests/svg/dom/id-reflect-expected.txt

This test checks that the id property on an SVGElement reflects the corresponding attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS g.id is ""
PASS g.id is "abc"
PASS g.getAttribute('id') is "def"
PASS tspan.id is ""
PASS tspan.id is "abc"
PASS tspan.getAttribute('id') is "def"
PASS foreignObject.id is ""
PASS foreignObject.id is "abc"
PASS foreignObject.getAttribute('id') is "def"
PASS successfullyParsed is true

TEST COMPLETE
